Make glfw and qt work by forcing x11
