+++ wgpu/gui/__init__.py
@@ -2,6 +2,7 @@
 Code to provide a canvas to render to.
 """
 
+from . import _gui_utils  # noqa: F401
 from .base import WgpuCanvasInterface, WgpuCanvasBase, WgpuAutoGui  # noqa: F401
 from .offscreen import WgpuOffscreenCanvasBase  # noqa: F401
