+++ wgpu/gui/base.py
@@ -1,10 +1,8 @@
-import os
 import sys
 import time
 import weakref
 import logging
 from contextlib import contextmanager
-import ctypes.util
 from collections import defaultdict
 
 from .._coreutils import error_message_hash
@@ -56,41 +54,6 @@ def weakbind(method):
 
     proxy.__name__ = class_func.__name__
     return proxy
-
-
-_x11_display = None
-
-
-def get_alt_x11_display():
-    """Get (the pointer to) a process-global x11 display instance."""
-    # Ideally we'd get the real display object used by the GUI toolkit.
-    # But this is not always possible. In that case, using an alt display
-    # object can be used.
-    global _x11_display
-    assert sys.platform.startswith("linux")
-    if _x11_display is None:
-        x11 = ctypes.CDLL(ctypes.util.find_library("X11"))
-        x11.XOpenDisplay.restype = ctypes.c_void_p
-        _x11_display = x11.XOpenDisplay(None)
-    return _x11_display
-
-
-_wayland_display = None
-
-
-def get_alt_wayland_display():
-    """Get (the pointer to) a process-global Wayland display instance."""
-    # Ideally we'd get the real display object used by the GUI toolkit.
-    # This creates a global object, similar to what we do for X11.
-    # Unfortunately, this segfaults, so it looks like the real display object
-    # is needed? Leaving this here for reference.
-    global _wayland_display
-    assert sys.platform.startswith("linux")
-    if _wayland_display is None:
-        wl = ctypes.CDLL(ctypes.util.find_library("wayland-client"))
-        wl.wl_display_connect.restype = ctypes.c_void_p
-        _wayland_display = w.wl_display_connect(None)
-    return _wayland_display
 
 
 class WgpuCanvasInterface:
